apps/backend/requirements.txt
fastapi
uvicorn 
pydantic
sqlalchemy
python-dotenv
pymysql
greenlet
alembic
firebase-admin
python-jose[cryptography]
redis
cryptography
httpx
python-multipart
websockets
pydantic[email]